save disk space on model checkpoints via "detached word embedding"
Add an option `--detached_embeddings` for loading word embeddings from a
separate file from the model checkpoint. This prevents the model word
embeddings from being saved to the model checkpoint, which is especially
useful when frozen word embeddings for large vocabularies are being
used.

This requires the embeddings to be loaded either via the `--new_vocab`
or the previously added `--src/tgt_embeddings` options in translation,
or the `--new_vocab` option in training from checkpoint.

For use with the `--src/tgt_embeddings` options, as well as for
later extraction form the model checkpoint, the embedding vectors
for special vocabulary words are still saved to the model checkpoint
(this feature is currently improperly implemented through saving the
first 32 vocabulary vectors and will be properly implemented in future
commits).

Furthermore, in order to save space do not save optimizer states when
they are not required, e.g. when the `--reset_optim all` option is set.